backend/requirements.txt
fastapi==0.104.0 
uvicorn==0.23.2 
sqlalchemy==2.0.22 
pymysql==1.1.0 
pydantic==2.4.2 
python-jose==3.3.0 
passlib==1.7.4 
python-multipart==0.0.6 
cryptography==41.0.4 
email-validator==2.0.0 
bcrypt==4.0.1 
